% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_hyperfile.R
\name{format_hyperfile}
\alias{format_hyperfile}
\title{Format hyperfile}
\arguments{
\item{hyperfile}{Choose the imported csv of the hyperfile of choice}

\item{indicator}{create a vector list of the indicators of interest}

\item{OU}{create a vector list of the operating units of interest}

\item{fiscal_year}{create a vector list of the fiscal years of interset (i.e. c("FY19", "FY20", ...))}
}
\value{
sorts the hyperfile with the indicator, OU, and FY of interest
}
\description{
A function to reduce the size of the Tableau hyperfile for quick recall
This function allows you to identify indicators or OUs of interest for figure development
}
