% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_achievement_fig.R
\name{target_achievement_fig}
\alias{target_achievement_fig}
\title{Target Achievement Figure}
\arguments{
\item{hyperfile}{select the hyperfile dataframe}

\item{indicator}{select the indicator of interest}

\item{OU}{create a vector list of the operating units of interest}

\item{fiscal_year}{create a vector list of the fiscal years of interset (i.e. c("FY19", "FY20", ...))}

\item{source}{write a string of the data source being used}
}
\value{
sorts the hyperfile with the indicator, OU, and FY of interest
}
\description{
A function to generate target achievement
}
